Add files via upload
This program demonstrates a working knowledge of cleaning techniques necessary for exploratory data analysis. Please bear in mind that the profile report generated will not be displayed by default in saved notebooks. I have placed dataset heads at the bottom of this file so that compliance with instruction may be observed.